Attempt to fix stream not closed properly when having exe input stream
Revert removal of cmd.wait

trying to make killsignal param work

updated version name

cleaning

+++ pkg/stdin/backchannel.go
@@ -45,6 +45,9 @@ func (c *Client) Stop() (err error) {
 	if c.cmd.Process == nil {
 		return nil
 	}
+	if c.OnClose != nil {
+		return c.OnClose()
+	}
 	return errors.Join(c.cmd.Process.Kill(), c.cmd.Wait())
 }
